+++ config/importmap.rb
@@ -1,5 +1,8 @@
 # frozen_string_literal: true
 
+pin "controllers/recording_studio_notifications_push/push_devices_controller",
+    preload: true
+
 pin_all_from RecordingStudioNotificationsPush::Engine.root.join("app/javascript/recording_studio_notifications_push/controllers"),
              under: "controllers/recording_studio_notifications_push",
              to: "recording_studio_notifications_push/controllers",